Get game type by name
